pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonify"
version = "1.0.0"
description = "A beautiful Spotify data visualizer that creates shareable insights from your listening habits"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sonify Team", email = "contact@sonify.app"}
]
maintainers = [
    {name = "Sonify Team", email = "contact@sonify.app"}
]
keywords = ["spotify", "music", "visualization", "data", "analysis", "charts", "graphs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Framework :: Flask",
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.3.3",
    "spotipy>=2.23.0",
    "python-dotenv>=1.0.0",
    "plotly>=5.17.0",
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    "requests>=2.31.0",
    "Werkzeug>=2.3.7",
    "Jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "safety>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/makalin/Sonify"
Documentation = "https://github.com/makalin/Sonify#readme"
Repository = "https://github.com/makalin/Sonify"
"Bug Tracker" = "https://github.com/makalin/Sonify/issues"

[project.scripts]
sonify = "run:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
] 